+++ analytics-core/src/test/java/com/rudder/analytics/messages/BatchTest.java
@@ -4,6 +4,11 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.rudder.analytics.TestUtils;
 import com.squareup.burst.BurstJUnit4;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -25,6 +30,33 @@ public class BatchTest {
 
     assertThat(batch.batch()).isEqualTo(messages);
     assertThat(batch.context()).isEqualTo(context);
+  }
+
+  @Test
+  public void createWithSentAt(TestUtils.MessageBuilderFactory factory) throws ParseException {
+    Message message =
+        factory
+            .get()
+            .userId("userId")
+            .sentAt(new SimpleDateFormat("yyyy-MM-dd").parse("2022-04-01"))
+            .build();
+    Map<String, String> context = ImmutableMap.of("foo", "bar");
+    List<Message> messages = ImmutableList.of(message);
+
+    Batch batch = Batch.create(context, messages);
+
+    assertThat(batch.sentAt()).isEqualTo(messages.get(0).sentAt());
+  }
+
+  @Test
+  public void createWithSentAtNull(TestUtils.MessageBuilderFactory factory) throws ParseException {
+    Message message = factory.get().userId("userId").sentAt(null).build();
+    Map<String, String> context = ImmutableMap.of("foo", "bar");
+    List<Message> messages = ImmutableList.of(message);
+
+    Batch batch = Batch.create(context, messages);
+
+    assertThat(batch.sentAt()).isEqualToIgnoringHours(new Date());
   }
 
   @Test