+++ analytics/src/main/java/com/rudder/analytics/internal/StopMessage.java
@@ -24,6 +24,12 @@ class StopMessage implements Message {
         throw new UnsupportedOperationException();
     }
 
+    @Nullable
+    @Override
+    public Date sentAt() {
+        throw new UnsupportedOperationException();
+    }
+
     @Nonnull
     @Override
     public Date timestamp() {